Enable systemd integration
podman is able to pass systemd notification sockets through to the
containerized application.
By enabling the systemd integration in icinga2 the host systemd is able
to see the runtime status of the containerized icinga2 instance,
especially for startup notifications.

libsystemd is already installed in the base image so no new runtime
dependency is introduced.
An explicit installation is added for more clarity.

Signed-off-by: Thomas Weißschuh <thomas@t-8ch.de>